backend/requirements.txt
Django==1.11.6
djangorestframework==3.7.0
djangorestframework-jwt==1.11.0
django-cors-headers==2.1.0
